Move the program enrolments API to the programs package
