--- client/packages/system/src/Patient/ProgramEnrolment/api/hooks/document/index.ts
@@ -1,7 +1,0 @@
-import { useInsertProgramEnrolment } from './useInsertProgramEnrolment';
-import { useUpdateProgramEnrolment } from './useUpdateProgramEnrolment';
-
-export const Document = {
-  useInsertProgramEnrolment,
-  useUpdateProgramEnrolment,
-};
