--- client/packages/system/src/Patient/ProgramEnrolment/api/hooks/index.ts
@@ -1,13 +1,0 @@
-import { Document } from './document';
-import { Utils } from './utils';
-
-export const useProgramEnrolment = {
-  utils: {
-    api: Utils.useProgramEnrolmentApi,
-  },
-
-  document: {
-    insert: Document.useInsertProgramEnrolment,
-    update: Document.useUpdateProgramEnrolment,
-  },
-};
