--- client/packages/system/src/Patient/ProgramEnrolment/api/hooks/utils/index.ts
@@ -1,5 +1,0 @@
-import { useProgramEnrolmentApi } from './useProgramEnrolmentApi';
-
-export const Utils = {
-  useProgramEnrolmentApi: useProgramEnrolmentApi,
-};
