--- client/packages/system/src/Patient/ProgramEnrolment/api/operations.graphql
@@ -1,67 +1,0 @@
-fragment ProgramEnrolmentDocument on DocumentNode {
-  id
-  name
-  parents
-  author
-  timestamp
-  type
-  data
-  documentRegistry {
-    uiSchemaType
-    documentType
-    context
-    formSchemaId
-    jsonSchema
-    uiSchema
-  }
-}
-
-fragment ProgramEnrolment on ProgramEnrolmentNode {
-  type
-  programPatientId
-  patientId
-  name
-  enrolmentDatetime
-  document {
-    ...ProgramEnrolmentDocument
-  }
-}
-
-query programEnrolmentById($storeId: String!, $programId: String!) {
-  programEnrolments(
-    storeId: $storeId
-    filter: { id: { equalTo: $programId } }
-  ) {
-    ... on ProgramEnrolmentConnector {
-      __typename
-      nodes {
-        ...ProgramEnrolment
-      }
-      totalCount
-    }
-  }
-}
-
-mutation insertProgramEnrolment(
-  $storeId: String!
-  $input: InsertProgramEnrolmentInput!
-) {
-  insertProgramEnrolment(storeId: $storeId, input: $input) {
-    ... on ProgramEnrolmentNode {
-      __typename
-      ...ProgramEnrolment
-    }
-  }
-}
-
-mutation updateProgramEnrolment(
-  $storeId: String!
-  $input: UpdateProgramEnrolmentInput!
-) {
-  updateProgramEnrolment(storeId: $storeId, input: $input) {
-    ... on ProgramEnrolmentNode {
-      __typename
-      ...ProgramEnrolment
-    }
-  }
-}
